total refactor
