+++ lib/ui/page/home/page/user/view.dart
@@ -15,23 +15,14 @@
 // along with this program. If not, see
 // <https://www.gnu.org/licenses/agpl-3.0.html>.
 
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import '/api/backend/schema.dart' show Presence;
 import '/domain/model/user.dart';
-import '/domain/repository/user.dart';
 import '/l10n/l10n.dart';
 import '/routes.dart';
-import '/themes.dart';
-import '/ui/page/call/widget/conditional_backdrop.dart';
-import '/ui/page/home/page/chat/message_field/view.dart';
 import '/ui/page/home/page/chat/widget/back_button.dart';
-import '/ui/page/home/page/my_profile/controller.dart';
-import '/ui/page/home/page/my_profile/widget/copyable.dart';
-import '/ui/page/home/page/my_profile/widget/field_button.dart';
+import '/ui/page/home/widget/actions.dart';
 import '/ui/page/home/widget/app_bar.dart';
 import '/ui/page/home/widget/avatar.dart';
 import '/ui/page/home/widget/block.dart';
@@ -41,8 +32,13 @@ import '/ui/widget/svg/svg.dart';
 import '/ui/widget/text_field.dart';
 import '/ui/widget/widget_button.dart';
 import '/util/message_popup.dart';
-import '/util/platform_utils.dart';
 import 'controller.dart';
+import 'widget/blocked.dart';
+import 'widget/blocked_field.dart';
+import 'widget/name.dart';
+import 'widget/num.dart';
+import 'widget/presence.dart';
+import 'widget/status.dart';
 
 /// View of the [Routes.user] page.
 class UserView extends StatelessWidget {
@@ -348,444 +344,5 @@ class UserView extends StatelessWidget {
     if (result == true) {
       await c.blacklist();
     }
-  }
-}
-
-/// Dense [Padding] wrapper.
-class _DenseWidget extends StatelessWidget {
-  const _DenseWidget(this.child);
-
-  /// [Widget] that will be wrapped in padding.
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-        padding: const EdgeInsets.fromLTRB(8, 4, 8, 4), child: child);
-  }
-}
-
-/// Basic [Padding] wrapper.
-class _PaddingWidget extends StatelessWidget {
-  const _PaddingWidget(this.child);
-
-  /// [Widget] that will be wrapped with padding.
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(padding: const EdgeInsets.all(8), child: child);
-  }
-}
-
-/// [Widget] which builds a stylized button representing a single action.
-class ActionWidget extends StatelessWidget {
-  const ActionWidget({
-    super.key,
-    this.text,
-    this.onPressed,
-    this.trailing,
-  });
-
-  /// Text to display in this [ActionWidget].
-  final String? text;
-
-  /// Trailing to display in this [ActionWidget].
-  final Widget? trailing;
-
-  /// Callback, called when this button is tapped or activated other way.
-  final void Function()? onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 8),
-      child: _DenseWidget(
-        FieldButton(
-          onPressed: onPressed,
-          text: text ?? '',
-          style: TextStyle(color: Theme.of(context).colorScheme.secondary),
-          trailing: trailing != null
-              ? Transform.translate(
-                  offset: const Offset(0, -1),
-                  child: Transform.scale(scale: 1.15, child: trailing),
-                )
-              : null,
-        ),
-      ),
-    );
-  }
-}
-
-/// [Widget] which returns the action buttons to do with this [User].
-class ActionsWidget extends StatelessWidget {
-  const ActionsWidget({
-    super.key,
-    required this.inContacts,
-    required this.inFavorites,
-    required this.status,
-    required this.blacklistStatus,
-    this.removeFromContacts,
-    this.hideChat,
-    this.clearChat,
-    this.blacklistUser,
-    this.user,
-    this.isBlacklisted,
-    this.addToContacts,
-    this.unfavoriteContact,
-    this.favoriteContact,
-    this.unmuteChat,
-    this.muteChat,
-    this.unblacklist,
-  });
-
-  /// Indicator whether this [user] is already in the contacts list of the
-  /// authenticated [MyUser].
-  final RxBool inContacts;
-
-  /// Indicator whether the [user] is favorite.
-  final RxBool inFavorites;
-
-  /// Status of the [user] fetching.
-  final Rx<RxStatus> status;
-
-  /// Status of a [blacklist] progression.
-  final Rx<RxStatus> blacklistStatus;
-
-  /// Reactive [User] itself.
-  final RxUser? user;
-
-  /// Indicator whether this [user] is blacklisted.
-  final BlacklistRecord? isBlacklisted;
-
-  /// Adds the [user] to the contacts list of the authenticated [MyUser].
-  final void Function()? addToContacts;
-
-  /// Removes the [user] from the favorites.
-  final void Function()? unfavoriteContact;
-
-  /// Marks the [user] as favorited.
-  final void Function()? favoriteContact;
-
-  /// Unmutes a [Chat]-dialog with the [user].
-  final void Function()? unmuteChat;
-
-  /// Mutes a [Chat]-dialog with the [user].
-  final void Function()? muteChat;
-
-  /// Removes the [user] from the blacklist of the authenticated [MyUser].
-  final void Function()? unblacklist;
-
-  /// Opens a confirmation popup deleting the [User] from address book.
-  final void Function()? removeFromContacts;
-
-  /// Opens a confirmation popup hiding the [Chat]-dialog with the [User].
-  final void Function()? hideChat;
-
-  /// Opens a confirmation popup clearing the [Chat]-dialog with the [User].
-  final void Function()? clearChat;
-
-  /// Opens a confirmation popup blacklisting the [User].
-  final void Function()? blacklistUser;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Obx(() {
-          return ActionWidget(
-            key: Key(inContacts.value
-                ? 'DeleteFromContactsButton'
-                : 'AddToContactsButton'),
-            text: inContacts.value
-                ? 'btn_delete_from_contacts'.l10n
-                : 'btn_add_to_contacts'.l10n,
-            onPressed: status.value.isLoadingMore
-                ? null
-                : inContacts.value
-                    ? () => removeFromContacts
-                    : addToContacts,
-          );
-        }),
-        Obx(() {
-          return ActionWidget(
-            text: inFavorites.value
-                ? 'btn_delete_from_favorites'.l10n
-                : 'btn_add_to_favorites'.l10n,
-            onPressed: inFavorites.value ? unfavoriteContact : favoriteContact,
-          );
-        }),
-        if (user?.user.value.dialog.isLocal == false &&
-            user?.dialog.value != null) ...[
-          Obx(() {
-            final chat = user!.dialog.value!.chat.value;
-            final bool isMuted = chat.muted != null;
-
-            return ActionWidget(
-              text: isMuted ? 'btn_unmute_chat'.l10n : 'btn_mute_chat'.l10n,
-              trailing: isMuted
-                  ? SvgImage.asset(
-                      'assets/icons/btn_mute.svg',
-                      width: 18.68,
-                      height: 15,
-                    )
-                  : SvgImage.asset(
-                      'assets/icons/btn_unmute.svg',
-                      width: 17.86,
-                      height: 15,
-                    ),
-              onPressed: isMuted ? unmuteChat : muteChat,
-            );
-          }),
-          ActionWidget(
-            text: 'btn_hide_chat'.l10n,
-            trailing: SvgImage.asset('assets/icons/delete.svg', height: 14),
-            onPressed: () => hideChat,
-          ),
-          ActionWidget(
-            key: const Key('ClearHistoryButton'),
-            text: 'btn_clear_history'.l10n,
-            trailing: SvgImage.asset('assets/icons/delete.svg', height: 14),
-            onPressed: () => clearChat,
-          ),
-        ],
-        ActionWidget(
-          key: Key(isBlacklisted != null ? 'Unblock' : 'Block'),
-          text: isBlacklisted != null ? 'btn_unblock'.l10n : 'btn_block'.l10n,
-          onPressed: isBlacklisted != null ? unblacklist : () => blacklistUser,
-          trailing: Obx(() {
-            return AnimatedOpacity(
-              duration: 200.milliseconds,
-              opacity: blacklistStatus.value.isEmpty ? 0 : 1,
-              child: const CustomProgressIndicator(),
-            );
-          }),
-        ),
-        ActionWidget(text: 'btn_report'.l10n, onPressed: () {}),
-      ],
-    );
-  }
-}
-
-/// [Widget] which returns a [User.name] copyable field.
-class NameWidget extends StatelessWidget {
-  const NameWidget({super.key, this.user});
-
-  /// Reactive [User] itself.
-  final RxUser? user;
-
-  @override
-  Widget build(BuildContext context) {
-    return _PaddingWidget(
-      CopyableTextField(
-        key: const Key('NameField'),
-        state: TextFieldState(
-          text: '${user?.user.value.name?.val ?? user?.user.value.num.val}',
-        ),
-        label: 'label_name'.l10n,
-        copy: '${user?.user.value.name?.val ?? user?.user.value.num.val}',
-      ),
-    );
-  }
-}
-
-/// [Widget] which returns a [User.status] copyable field.
-class StatusWidget extends StatelessWidget {
-  const StatusWidget({super.key, this.user});
-
-  /// Reactive [User] itself.
-  final RxUser? user;
-
-  @override
-  Widget build(BuildContext context) {
-    return Obx(() {
-      final UserTextStatus? status = user?.user.value.status;
-
-      if (status == null) {
-        return Container();
-      }
-
-      return _PaddingWidget(
-        CopyableTextField(
-          key: const Key('StatusField'),
-          state: TextFieldState(text: status.val),
-          label: 'label_status'.l10n,
-          copy: status.val,
-        ),
-      );
-    });
-  }
-}
-
-/// [Widget] which returns a [User.presence] text.
-class PresenceWidget extends StatelessWidget {
-  const PresenceWidget({super.key, this.user});
-
-  /// Reactive [User] itself.
-  final RxUser? user;
-
-  @override
-  Widget build(BuildContext context) {
-    return Obx(() {
-      final Presence? presence = user?.user.value.presence;
-      if (presence == null) {
-        return Container();
-      }
-
-      final subtitle = user?.user.value.getStatus();
-
-      return _PaddingWidget(
-        ReactiveTextField(
-          key: const Key('Presence'),
-          state: TextFieldState(text: subtitle),
-          label: 'label_presence'.l10n,
-          enabled: false,
-          trailing: CircleAvatar(
-            key: Key(presence.name.capitalizeFirst!),
-            backgroundColor: presence.getColor(),
-            radius: 7,
-          ),
-        ),
-      );
-    });
-  }
-}
-
-/// [Widget] which returns a [User.num] copyable field.
-class NumWidget extends StatelessWidget {
-  const NumWidget({super.key, this.user});
-
-  /// Reactive [User] itself.
-  final RxUser? user;
-
-  @override
-  Widget build(BuildContext context) {
-    return _PaddingWidget(
-      CopyableTextField(
-        key: const Key('UserNum'),
-        state: TextFieldState(
-          text: user!.user.value.num.val.replaceAllMapped(
-            RegExp(r'.{4}'),
-            (match) => '${match.group(0)} ',
-          ),
-        ),
-        label: 'label_num'.l10n,
-        copy: user?.user.value.num.val,
-      ),
-    );
-  }
-}
-
-/// [Widget] which returns the blacklisted information of this [User].
-class BlockedWidget extends StatelessWidget {
-  const BlockedWidget({super.key, this.isBlacklisted});
-
-  /// Indicates whether [User] is blacklisted.
-  final BlacklistRecord? isBlacklisted;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        if (isBlacklisted?.at != null)
-          _PaddingWidget(
-            ReactiveTextField(
-              state: TextFieldState(text: isBlacklisted!.at.toString()),
-              label: 'label_date'.l10n,
-              enabled: false,
-            ),
-          ),
-        if (isBlacklisted?.reason != null)
-          _PaddingWidget(
-            ReactiveTextField(
-              state: TextFieldState(text: isBlacklisted!.reason?.val),
-              label: 'label_reason'.l10n,
-              enabled: false,
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-/// [Widget] which returns a [WidgetButton] for removing the [User] from the
-/// blacklist.
-class BlockedField extends StatelessWidget {
-  const BlockedField({super.key, required this.unblacklist});
-
-  /// Removes the [User] from the blacklist of the authenticated [RxUser].
-  final Future<void> Function() unblacklist;
-
-  @override
-  Widget build(BuildContext context) {
-    final Style style = Theme.of(context).extension<Style>()!;
-
-    return Theme(
-      data: MessageFieldView.theme(context),
-      child: SafeArea(
-        child: Container(
-          key: const Key('BlockedField'),
-          decoration: BoxDecoration(
-            borderRadius: style.cardRadius,
-            boxShadow: const [
-              CustomBoxShadow(
-                blurRadius: 8,
-                color: Color(0x22000000),
-              ),
-            ],
-          ),
-          child: ConditionalBackdropFilter(
-            condition: style.cardBlur > 0,
-            filter: ImageFilter.blur(
-              sigmaX: style.cardBlur,
-              sigmaY: style.cardBlur,
-            ),
-            borderRadius: style.cardRadius,
-            child: Container(
-              constraints: const BoxConstraints(minHeight: 56),
-              decoration: BoxDecoration(color: style.cardColor),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.end,
-                children: [
-                  Expanded(
-                    child: Padding(
-                      padding: EdgeInsets.only(
-                        top: 5 + (PlatformUtils.isMobile ? 0 : 8),
-                        bottom: 13,
-                      ),
-                      child: Transform.translate(
-                        offset: Offset(0, PlatformUtils.isMobile ? 6 : 1),
-                        child: WidgetButton(
-                          onPressed: unblacklist,
-                          child: IgnorePointer(
-                            child: ReactiveTextField(
-                              enabled: false,
-                              key: const Key('MessageField'),
-                              state: TextFieldState(text: 'btn_unblock'.l10n),
-                              filled: false,
-                              dense: true,
-                              textAlign: TextAlign.center,
-                              padding: const EdgeInsets.symmetric(vertical: 8),
-                              style: style.boldBody.copyWith(
-                                fontSize: 17,
-                                color: Theme.of(context).colorScheme.secondary,
-                              ),
-                              type: TextInputType.multiline,
-                              textInputAction: TextInputAction.newline,
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
   }
 }
